requirements.txt
streamlit
librosa
openai-whisper
spacy
plotly
pandas
numpy
soundfile
imageio-ffmpeg
langchain
langchain-openai
python-dotenv
